Add Final touches on Tools section
We made the final color/ font changes to the css file so that all of the
classes that were required to have different looks, have different
looks.  We made it so that different classes links will change colors
when hovered over, and changed some of the fonts as well.

co-authored-by morrisschoolsjm17